GEOMESA-1575 GeoMesa Spark module decompose and refactor
